+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Changelog for NeoFS Node
 - Split object middle children treated as root objects in SearchV2 (#3316)
 - Broken "owner" parameter of "accounting balance" CLI command (#3325)
 - Unclosed network connections in CLI (#3326)
+- SN no longer fails tombstone verification on `ALREADY_REMOVED` member's status (#3327)
 
 ### Changed
 - IR calls `ObjectService.SearchV2` to select SG objects now (#3144)